Some minor fixes to GPS (Important notes below)
There is a problem with the GPS permission. When the player opens the game, the game will ask for permission. However, even if the player allows it, the game will not generate the map the first time (probably because the scripts ran before the permission).

To fix this, I could run the scripts again, but there are dependencies and I'd rather not get into them due to time constraints.

Another option (prefered for now), is to check if the player has the GPS on in the Main Menu Scene. Doing that allows the player to restart the game or enable the GPS before proceeding.

How the game goes in general:
(first time only - request gps) -> Main Menu -> Lobby -> Game World